fix : 제안서 pages 폴더 정리
