+++ src/App.js
@@ -5,20 +5,21 @@ import LoginHome from "./pages/Auth/LoginHome";
 import MainHome from "./pages/Main/MainHome";
 import ShopDetailPage from "./pages/ShopDetailPage";
 import AddSchedulePage from "./pages/AddSchedulePage";
-import CreateProposal from "./pages/CreateProposal";
 
 import SearchPage from "./pages/ShopSearch/SearchPage";
 import Recommend from "./pages/ShopSearch/Recommend";
 import OurService from "./pages/OurService";
 import ShopMemberMyPage from "./pages/MyPage/ShopMemberMyPage";
-import AllProposal from "./pages/AllProposal";
+
+// 제안서 관련
+import Proposal from "./pages/Proposal/Proposal";
+import AllProposal from "./pages/Proposal/AllProposal";
+import EditProposal from "./pages/EditProposal/EditProposal";
+import CreateProposal from "./pages/CreateProposal/CreateProposal";
 
 import ShopMypageStart from "./pages/ShopMyPageStart";
-
 import ShopMyPageProgressive from "./pages/ShopMyPageProgressive";
 import PickupSchedulePage from "./pages/PickupSchedulePage";
-import EditProposal from "./pages/EditProposal";
-
 import ShopInformationRegister from "./pages/MyPage/ShopInformationRegister";
 import ShopInformationModify from "./pages/MyPage/ShopInformationModify";
 import LoginLoading from "./pages/Auth/LoginLoading";
@@ -28,9 +29,10 @@ import TraineeRoute from "./route/TraineeRoute";
 import BakerRoute from "./route/BakerRoute";
 import AccountSetting from "./pages/AccountSetting";
 import ShopCakerMyPage from "./pages/MyPage/ShopCakerMyPage";
-import Proposal from "./pages/Proposal";
 
 import Kakao from "./pages/Auth/Kakao";
+import NoPermission from "./pages/Auth/NoPermission";
+
 function App() {
   console.log(userRole);
   return (
@@ -84,6 +86,8 @@ function App() {
           path="/shop/addschedule/:storeId"
           element={<AddSchedulePage />}
         />
+        {/* 잘못된 접근 */}
+        <Route exact path="/accessfail" element={<NoPermission />} />
       </Routes>
     </BrowserRouter>
   );
