+++ src/components/Proposal/CustomCalendar.js
@@ -21,44 +21,7 @@ const Right = styled.img`
   margin-right: 13px;
 `;
 
-
 const CustomCalendar = ({ setClickedDay, original, setOriginal }) => {
-  // 삭제하면 오류나서... 일단 여기에 옮겨둘게요
-  const [existArr, setExistArr] = useState([
-    { id: 0, exist: false, reservation: "예약 정보" },
-    { id: 1, exist: false, reservation: "예약 정보" },
-    { id: 2, exist: false, reservation: "예약 정보" },
-    { id: 3, exist: false, reservation: "예약 정보" },
-    { id: 4, exist: false, reservation: "예약 정보" },
-    { id: 5, exist: false, reservation: "예약 정보" },
-    { id: 6, exist: false, reservation: "예약 정보" },
-    { id: 7, exist: false, reservation: "예약 정보" },
-    { id: 8, exist: false, reservation: "예약 정보" },
-    { id: 9, exist: false, reservation: "예약 정보" },
-    { id: 10, exist: false, reservation: "예약 정보" },
-    { id: 11, exist: false, reservation: "예약 정보" },
-    { id: 12, exist: false, reservation: "예약 정보" },
-    { id: 13, exist: false, reservation: "예약 정보" },
-    { id: 14, exist: false, reservation: "예약 정보" },
-    { id: 15, exist: false, reservation: "예약 정보" },
-    { id: 16, exist: false, reservation: "예약 정보" },
-    { id: 17, exist: false, reservation: "예약 정보" },
-    { id: 18, exist: false, reservation: "예약 정보" },
-    { id: 19, exist: false, reservation: "예약 정보" },
-    { id: 20, exist: false, reservation: "예약 정보" },
-    { id: 21, exist: false, reservation: "예약 정보" },
-    { id: 22, exist: false, reservation: "예약 정보" },
-    { id: 23, exist: false, reservation: "예약 정보" },
-    { id: 24, exist: false, reservation: "예약 정보" },
-    { id: 25, exist: false, reservation: "예약 정보" },
-    { id: 26, exist: false, reservation: "예약 정보" },
-    { id: 27, exist: false, reservation: "예약 정보" },
-    { id: 28, exist: false, reservation: "예약 정보" },
-    { id: 29, exist: false, reservation: "예약 정보" },
-    { id: 30, exist: false, reservation: "예약 정보" },
-    { id: 31, exist: false, reservation: "예약 정보" },
-  ]);
-
   const today = {
     year: new Date().getFullYear(), //오늘 연도
     month: new Date().getMonth() + 1, //오늘 월
@@ -158,19 +121,19 @@ const CustomCalendar = ({ setClickedDay, original, setOriginal }) => {
   // allDaysHavingSchedule은
   // { year: , month: , date: }인 객체를 가지는 배열
   // 해당 파라미터 생략하여 캘린더 사용하면 아래 조건문 실행 X
-  if (allDaysHavingSchedule) {
-    const dateHavingSchedule = allDaysHavingSchedule
-      .filter(
-        day => day["year"] === selectedYear && day["month"] === selectedMonth,
-      )
-      .map(day => day["date"]);
-    existArr = existArr.map(exist => {
-      return {
-        ...exist,
-        exist: dateHavingSchedule.includes(exist.id),
-      };
-    });
-  }
+  // if (allDaysHavingSchedule) {
+  //   const dateHavingSchedule = allDaysHavingSchedule
+  //     .filter(
+  //       day => day["year"] === selectedYear && day["month"] === selectedMonth,
+  //     )
+  //     .map(day => day["date"]);
+  //   existArr = existArr.map(exist => {
+  //     return {
+  //       ...exist,
+  //       exist: dateHavingSchedule.includes(exist.id),
+  //     };
+  //   });
+  // }
   const returnDay = () => {
     //선택된 달의 날짜들 반환 함수
     let dayArr = [];
