+++ src/pages/EditProposal/EditProposal.js
@@ -2,11 +2,11 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import { Route, Routes, useNavigate, useLocation } from "react-router-dom";
 
-import TopBar from "../components/Common/Sidebar/TopBar";
-import PickUp from "./EditProposal/PickUp";
-import Done from "./EditProposal/Done";
+import TopBar from "../../components/Common/Sidebar/TopBar";
+import PickUp from "./PickUp";
+import Done from "./Done";
 
-import http from "../common/http";
+import http from "../../common/http";
 
 const EditProposal = () => {
   const [data, setData] = useState(null);
